+++ sktime/transformations/series/compose.py
@@ -8,9 +8,6 @@ from sklearn.base import clone
 from sklearn.utils.metaestimators import if_delegate_has_method
 
 from sktime.transformations.base import BaseTransformer
-from sktime.transformations.series.exponent import (
-    ExponentTransformer as _ExponentTransformer,
-)
 from sktime.utils.validation.series import check_series
 
 __author__ = ["aiwalter", "SveaMeyer13"]
@@ -433,145 +430,3 @@ def _check_is_pdseries(z):
         z = z.to_frame()
         is_series = True
     return z, is_series
-
-
-class FitInTransform(BaseTransformer):
-    """Transformer compostion to always fit a given transformer on the transform data only.
-
-    Some transformers like Imputer(method="mean") might be more accurate for some use
-    cases if they are fitted only on the transform/predict data. The FitInTransform
-    transformer can be best used in a pipeline context where train data is different to
-    the transform/predict data.
-
-    Parameters
-    ----------
-    transformer : Estimator
-        scikit-learn-like or sktime-like transformer to fit and apply to series.
-
-    Attributes
-    ----------
-    transformer_ : Fitted transformer.
-
-    Examples
-    --------
-    >>> from sktime.datasets import load_longley
-    >>> from sktime.forecasting.naive import NaiveForecaster
-    >>> from sktime.forecasting.base import ForecastingHorizon
-    >>> from sktime.forecasting.compose import ForecastingPipeline
-    >>> from sktime.forecasting.model_selection import temporal_train_test_split
-    >>> from sktime.transformations.series.compose import FitInTransform
-    >>> from sktime.transformations.series.impute import Imputer
-
-    >>> y, X = load_longley()
-    >>> y_train, y_test, X_train, X_test = temporal_train_test_split(y, X)
-    >>> fh = ForecastingHorizon(y_test.index, is_relative=False)
-
-    >>> # we want to fit the Imputer only on the predict (=transform) data.
-    >>> # note that NaiveForecaster cant use X data, this is just a show case.
-    >>> pipe = ForecastingPipeline(
-    ...     steps=[
-    ...         ("imputer", FitInTransform(Imputer(method="mean"))),
-    ...         ("forecaster", NaiveForecaster()),
-    ...     ]
-    ... )
-    >>> pipe.fit(y_train, X_train)
-    ForecastingPipeline(...)
-    >>> y_pred = pipe.predict(fh=fh, X=X_test)
-    """
-
-    def __init__(self, transformer):
-        self.transformer = transformer
-        super(FitInTransform, self).__init__()
-        self.clone_tags(transformer, None)
-        self.set_tags(**{"fit_is_empty": True})
-        # self.set_tags(handles-missing-data=True)
-
-    def _transform(self, X, y=None):
-        """Transform X and return a transformed version.
-
-        private _transform containing core logic, called from transform
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _transform must support all types in it
-            Data to be transformed
-        y : Series or Panel of mtype y_inner_mtype, default=None
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        transformed version of X
-        """
-        self.transformer_ = clone(self.transformer)
-        self.transformer_.fit(X=X, y=y)
-        return self.transformer_.transform(X=X, y=y)
-
-    def _inverse_transform(self, X, y=None):
-        """Inverse transform, inverse operation to transform.
-
-        private _inverse_transform containing core logic, called from inverse_transform
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _inverse_transform must support all types in it
-            Data to be inverse transformed
-        y : Series or Panel of mtype y_inner_mtype, optional (default=None)
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        inverse transformed version of X
-        """
-        return self.transformer_.inverse_transform(X=X, y=y)
-
-    def _update(self, X, y=None):
-        """Update transformer with X and y.
-
-        private _update containing the core logic, called from update
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _update must support all types in it
-            Data to update transformer with
-        y : Series or Panel of mtype y_inner_mtype, default=None
-            Additional data, e.g., labels for tarnsformation
-
-        Returns
-        -------
-        self: reference to self
-        """
-        self.transformer_.update(X=X, y=y)
-
-    def get_fitted_params(self):
-        """Get fitted parameters.
-
-        Returns
-        -------
-        fitted_params : dict
-        """
-        return self.transformer_.get_fitted_params()
-
-    @classmethod
-    def get_test_params(cls, parameter_set="default"):
-        """Return testing parameter settings for the estimator.
-
-        Parameters
-        ----------
-        parameter_set : str, default="default"
-            Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return `"default"` set.
-            There are currently no reserved values for transformers.
-
-        Returns
-        -------
-        params : dict or list of dict, default = {}
-            Parameters to create testing instances of the class
-            Each dict are parameters to construct an "interesting" test instance, i.e.,
-            `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
-            `create_test_instance` uses the first (or only) dictionary in `params`
-        """
-        params = {"transformer": _ExponentTransformer()}
-        return params
